quiz15

+++ quizzes/quiz15/quiz15.c
@@ -4,10 +4,9 @@ long* SumIntPairsToLongs(int ints[], size_t numOfElements)
 {
 	size_t i = 0;
 	long* longs = (long*)ints;
-	
 	for(i = 0; i < numOfElements / 2; ++i)
 	{
-		longs[i] = ints[i*2] + ints[i*2 + 1];
+		longs[i] = (long)ints[i*2] + ints[i*2 + 1];
 	}
 	return longs;
 }
